+++ CHANGES.md
@@ -36,6 +36,7 @@
   EXIF `Orientation` tag.
 * KakaduNativeProcessor and OpenJpegProcessor support IPTC and XMP metadata.
 * TurboJpegProcessor supports EXIF, IPTC, and XMP metadata.
+* Java2dProcessor supports the XPM source format.
 * FfmpegProcessor, KakaduNativeProcessor, OpenJpegProcessor, and
   PdfBoxProcessor use libjpeg-turbo to write JPEGs if it is available.
 * Added the `scaled` position for image overlays.